Added Checks before AddingNewFavourite. AddFavourite returns FavouriteReadDTO instead of string as requested. Modified RemoveCommandHandler as requested.

+++ Eskon.Core/Features/FavouriteFeatures/Commands/Handler/FavouriteCommandHandler.cs
@@ -2,45 +2,56 @@
 using Eskon.Core.Features.FavouriteFeatures.Commands.Command;
 using Eskon.Core.Features.FavouriteFeatures.Commands.Handler;
 using Eskon.Core.Response;
-using Eskon.Domian.Entities.Identity;
+using Eskon.Domian.DTOs.Favourite;
 using Eskon.Domian.Models;
 using Eskon.Service.UnitOfWork;
-using Microsoft.AspNetCore.Identity;
 
 
 namespace Eskon.Core.Features.PropertyFeatures.Commands.Handler
 {
-    public class PropertyCommandHandler : ResponseHandler, IFavouriteCommandHandler
+    public class FavouriteCommandHandler : ResponseHandler, IFavouriteCommandHandler
     {
         #region Fields
         private readonly IServiceUnitOfWork _serviceUnitOfWork;
         private readonly IMapper _mapper;
         #endregion
-        public PropertyCommandHandler(IMapper mapper, IServiceUnitOfWork serviceUnitOfWork, UserManager<User> userManager)
+        public FavouriteCommandHandler(IMapper mapper, IServiceUnitOfWork serviceUnitOfWork)
         {
             _serviceUnitOfWork = serviceUnitOfWork;
             _mapper = mapper;
         }
 
-        public async Task<Response<string>> Handle(AddFavouriteCommand request, CancellationToken cancellationToken)
+        public async Task<Response<FavouriteReadDTO>> Handle(AddFavouriteCommand request, CancellationToken cancellationToken)
         {
-            var favourite = new Favourite
-            {
-                UserId = request.UserId,
-                PropertyId = request.PropertyId
-            }; await _serviceUnitOfWork.FavouriteService.AddFavouriteAsync(favourite);
-            await _serviceUnitOfWork.SaveChangesAsync();
+            var property = await _serviceUnitOfWork.PropertyService.GetPropertyByIdAsync(request.PropertyId);
+            if (property == null)
+                return NotFound<FavouriteReadDTO>("Property Not Found");
 
-           
-            return Created("Added to Favourites Successfully");
-        }
-        public async Task<Response<string>> Handle(RemoveFavouriteCommand request, CancellationToken cancellationToken)
-        {
-            var favourite = new Favourite
+            var favourite = await _serviceUnitOfWork.FavouriteService.GetFavouriteByIdAsync(request.UserId, request.PropertyId);
+            if (favourite != null)
+                return BadRequest<FavouriteReadDTO>("This Property is already on your favourites list");
+
+            favourite = new Favourite
             {
                 UserId = request.UserId,
                 PropertyId = request.PropertyId
             };
+
+            await _serviceUnitOfWork.FavouriteService.AddFavouriteAsync(favourite);
+            await _serviceUnitOfWork.SaveChangesAsync();
+            favourite = await _serviceUnitOfWork.FavouriteService.GetFavouriteByIdAsync(request.UserId, request.PropertyId);
+            var favouriteDTO = _mapper.Map<FavouriteReadDTO>(favourite); 
+
+            return Created(favouriteDTO,"Property added to Favourites Succefully");
+        }
+
+        public async Task<Response<string>> Handle(RemoveFavouriteCommand request, CancellationToken cancellationToken)
+        {
+            var favourite = await _serviceUnitOfWork.FavouriteService.GetFavouriteByIdAsync(request.UserId, request.PropertyId);
+            if (favourite == null)
+            {
+                return NotFound<string>($"Property Not Found in favourites");
+            }
             await _serviceUnitOfWork.FavouriteService.RemoveFavouriteAsync(favourite);
             await _serviceUnitOfWork.SaveChangesAsync();
